docs(rate): update docs and demo
