+++ src/rate/__test__/demo.test.jsx
@@ -4,23 +4,25 @@
 
 import { mount } from '@vue/test-utils';
 import allowHalfVue from '@/rate/demos/allow-half.vue';
-import baseVue from '@/rate/demos/base.vue';
 import colorVue from '@/rate/demos/color.vue';
 import countVue from '@/rate/demos/count.vue';
 import disabledVue from '@/rate/demos/disabled.vue';
+import filledVue from '@/rate/demos/filled.vue';
 import iconVue from '@/rate/demos/icon.vue';
 import mobileVue from '@/rate/demos/mobile.vue';
+import outlineVue from '@/rate/demos/outline.vue';
 import sizeVue from '@/rate/demos/size.vue';
 import textVue from '@/rate/demos/text.vue';
 
 const mapper = {
   allowHalfVue,
-  baseVue,
   colorVue,
   countVue,
   disabledVue,
+  filledVue,
   iconVue,
   mobileVue,
+  outlineVue,
   sizeVue,
   textVue,
 };
