+++ src/rate/demos/allow-half.vue
@@ -1,5 +1,8 @@
 <template>
-  <t-rate v-model="value" clearable allow-half :gap="defaultGap" />
+  <div class="rate-demo-cell">
+    <div class="rate-demo-cell__label">半星评价</div>
+    <t-rate v-model="value" clearable allow-half :gap="defaultGap" />
+  </div>
 </template>
 
 <script lang="ts" setup>
