+++ src/rate/demos/count.vue
@@ -1,5 +1,8 @@
 <template>
-  <t-rate v-model="value" :count="6" clearable :gap="defaultGap" />
+  <div class="rate-demo-cell">
+    <div class="rate-demo-cell__label">自定义数量</div>
+    <t-rate v-model="value" :count="4" clearable :gap="defaultGap" />
+  </div>
 </template>
 
 <script lang="ts" setup>
