+++ src/rate/demos/disabled.vue
@@ -1,5 +1,8 @@
 <template>
-  <t-rate v-model="value" disabled :gap="defaultGap" />
+  <div class="rate-demo-cell">
+    <div class="rate-demo-cell__label">仅展示</div>
+    <t-rate v-model="value" disabled :gap="defaultGap" />
+  </div>
 </template>
 
 <script lang="ts" setup>
