+++ src/rate/demos/icon.vue
@@ -1,9 +1,12 @@
 <template>
-  <t-rate v-model="value" clearable allow-half>
-    <template #icon>
-      <CheckCircleFilledIcon />
-    </template>
-  </t-rate>
+  <div class="rate-demo-cell">
+    <div class="rate-demo-cell__label">规格20</div>
+    <t-rate v-model="value" clearable allow-half>
+      <template #icon>
+        <CheckCircleFilledIcon />
+      </template>
+    </t-rate>
+  </div>
 </template>
 
 <script lang="ts" setup>
