+++ src/rate/demos/outline.vue
@@ -1,0 +1,6 @@
+<template>
+  <div class="rate-demo-cell">
+    <div class="rate-demo-cell__label">请点击评分</div>
+    <t-rate :default-value="3" variant="outline" clearable :gap="8" />
+  </div>
+</template>
